+++ substratevm/src/com.oracle.svm.core/src/com/oracle/svm/core/graal/code/SubstrateCompilationIdentifier.java
@@ -28,14 +28,13 @@ import java.util.concurrent.atomic.AtomicLong;
 
 import jdk.graal.compiler.core.common.CompilationIdentifier;
 import jdk.graal.compiler.debug.GraalError;
-
 import jdk.vm.ci.meta.ResolvedJavaMethod;
 
 public class SubstrateCompilationIdentifier implements CompilationIdentifier {
 
     private static final AtomicLong uniqueIds = new AtomicLong();
 
-    private final long id;
+    protected final long id;
     private final ResolvedJavaMethod method;
 
     public SubstrateCompilationIdentifier(ResolvedJavaMethod method) {